# docs/swagger.yaml
definitions:
  controller.response:
    properties:
      message:
        type: string
    type: object
...
  models.Admins:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      lastVisit:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  models.Applicant:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      resume:
        type: string
      status:
        type: integer
      vacancy:
        $ref: '#/definitions/models.Vacancy'
      vacancyId:
        type: integer
    type: object
...
  models.Vacancy:
    properties:
      created:
        $ref: '#/definitions/models.Admins'
      createdAt:
        type: string
      descriptionEn:
        type: string
      descriptionRu:
        type: string
      descriptionUz:
        type: string
      id:
        type: integer
      image:
        type: string
      isActive:
        type: boolean
      isDeleted:
        type: boolean
      nameEn:
        type: string
      nameRu:
        type: string
      nameUz:
        type: string
      region:
        type: string
      requirementEn:
        type: string
      requirementRu:
        type: string
      requirementUz:
        type: string
      responsibilityEn:
        type: string
      responsibilityRu:
        type: string
      responsibilityUz:
        type: string
      typeEn:
        type: string
      typeRu:
        type: string
      typeUz:
        type: string
      updated:
        $ref: '#/definitions/models.Admins'
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server CRM server.
  title: Golang CRM Swagger Documentation
  version: "1.0"
paths:
...
  /api/vacancy:
    get:
      consumes:
      - application/json
      description: this api is get vacancy
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Vacancy'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Get vacancy
      tags:
      - Vacancy
    post:
      consumes:
      - application/json
      description: this api is create vacancy
      parameters:
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: isActive
        type: boolean
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: region
        type: string
      - in: formData
        name: requirementEn
        type: string
      - in: formData
        name: requirementRu
        type: string
      - in: formData
        name: requirementUz
        type: string
      - in: formData
        name: responsibilityEn
        type: string
      - in: formData
        name: responsibilityRu
        type: string
      - in: formData
        name: responsibilityUz
        type: string
      - in: formData
        name: typeEn
        type: string
      - in: formData
        name: typeRu
        type: string
      - in: formData
        name: typeUz
        type: string
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create vacancy
      tags:
      - Vacancy
  /api/vacancy/{id}:
    delete:
      consumes:
      - application/json
      description: this api is DELETE vacancy
      parameters:
      - description: vacancy id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: DELETE vacancy
      tags:
      - Vacancy
    get:
      consumes:
      - application/json
      description: this api is Update vacancy
      parameters:
      - description: vacancy id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Update vacancy
      tags:
      - Vacancy
    put:
      consumes:
      - application/json
      description: this api is Update vacancy
      parameters:
      - description: vacancy id
        in: path
        name: id
        required: true
        type: string
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: isActive
        type: boolean
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: region
        type: string
      - in: formData
        name: requirementEn
        type: string
      - in: formData
        name: requirementRu
        type: string
      - in: formData
        name: requirementUz
        type: string
      - in: formData
        name: responsibilityEn
        type: string
      - in: formData
        name: responsibilityRu
        type: string
      - in: formData
        name: responsibilityUz
        type: string
      - in: formData
        name: typeEn
        type: string
      - in: formData
        name: typeRu
        type: string
      - in: formData
        name: typeUz
        type: string
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Update vacancy
      tags:
      - Vacancy
  /api/vacancy/all:
    get:
      consumes:
      - application/json
      description: this api is get all vacancy
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Vacancy'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Get all vacancy
      tags:
      - Vacancy
  /api/vacancy/applicant:
    get:
      consumes:
      - application/json
      description: this api is get applicant of vacancy statuses(0--new , 1-accept,2-saved,3-rejected)
      parameters:
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: vacancyId
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Applicant'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create applicant to vacancy
      tags:
      - Vacancy
    post:
      consumes:
      - application/json
      description: this api is create applicant to vacancy
      parameters:
      - in: formData
        name: description
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: phone
        type: string
      - in: formData
        name: vacancyId
        type: integer
      - description: file
        in: formData
        name: resume_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Create applicant to vacancy
      tags:
      - Vacancy
  /api/vacancy/applicant/{id}:
    delete:
      consumes:
      - application/json
      description: this api is create applicant to vacancy
      parameters:
      - description: vacancy id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create applicant to vacancy
      tags:
      - Vacancy
    get:
      consumes:
      - application/json
      description: this api is create applicant to vacancy
      parameters:
      - description: vacancy id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Applicant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create applicant to vacancy
      tags:
      - Vacancy
  /api/vacancy/applicant/accept/{id}:
    put:
      consumes:
      - application/json
      description: this api is for change status applicant to 1
      parameters:
      - description: applicant id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Applicant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Accept applicant
      tags:
      - Vacancy
  /api/vacancy/applicant/reject/{id}:
    put:
      consumes:
      - application/json
      description: this api is for change status applicant to 3
      parameters:
      - description: applicant id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Applicant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Reject applicant
      tags:
      - Vacancy
  /api/vacancy/applicant/save/{id}:
    put:
      consumes:
      - application/json
      description: this api is change applicant status to 2
      parameters:
      - description: applicant id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Applicant'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Save applicant
      tags:
      - Vacancy
securityDefinitions:
  BearerAuth:
    description: AUTH.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"